Update order support to include ActiveRecord::Relation(s) in addition to the base model class
