+++ lib/modern_searchlogic/column_conditions.rb
@@ -29,19 +29,6 @@ module ModernSearchlogic
           self.method(method).arity
       end
 
-      def order(expression)
-        match_data = expression.to_s.match(/^(ascend|descend)_by_(.*)/)
-        return super unless match_data
-
-        direction = match_data.captures.first
-        column = match_data.captures.second
-        if direction == 'ascend'
-          order(arel_table[column].asc)
-        else
-          order(arel_table[column].desc)
-        end
-      end
-
       private
 
       def searchlogic_scope_dynamically_defined?(method)
