+++ lib/modern_searchlogic/railtie.rb
@@ -2,7 +2,7 @@ module ModernSearchlogic
   class Railtie < Rails::Railtie
     initializer 'modern_searchlogic.setup_activerecord' do
       ActiveSupport.on_load(:active_record) do
-        ActiveRecord::Base.__send__(:include, ModernSearchlogic::ActiveRecordMethods)
+        ModernSearchlogic::ActiveRecordMethods.install
       end
     end
   end
